mini_trello_be/requirements.txt
Django>=3.2,<4.0
boto3
graphene-django>=2.15.0
django-cors-headers>=3.7.0
djangorestframework>=3.12.4
django-graphql-jwt